Make the relationship between supply and price non-linear.
When this relationship was linear, any change in supply or demand would
be exactly compensated by a change in price.  I implemented an inverse
(i.e. 1/x) relationship to fix this.  There are many other relationships
I could've implemented, but I chose this one due to its simplicity and
its flexibility.